requirements.txt
sentence-transformers>=2.2.2
faiss-cpu>=1.7.3.post2
transformers>=4.30.0
torch>=2.0.0
PyPDF2>=3.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
pdfplumber>=0.7.6